chore: move external repositories to persistence package (#1)
